Добавил возможность выбирать топпинг
и к тому же еще поменял тип возврашения с  double на int

+++ SDProject/src/com/company/entities/Goulash.java
@@ -14,7 +14,7 @@ public class Goulash extends Dishes {
     }
 
     @Override
-    public double cost() {
+    public int cost() {
         return 450;
     }
