+++ SDProject/src/com/company/entities/Manty.java
@@ -14,7 +14,7 @@ public class Manty extends Dishes {
     }
 
     @Override
-    public double cost() {
+    public int cost() {
         return 300;
     }
